requirement.txt
pandas_ta
stable_baselines3
shimmy
gymnasium